Big set of changes to prepare for client-side prediction.

+++ src/main/java/edu/stanford/junction/props/HistoryMAC.java
@@ -1,9 +1,9 @@
 package edu.stanford.junction.props;
 
 class HistoryMAC{
-	public int stateNumber;
+	public long stateNumber;
 	public String stateHash;
-	public HistoryMAC(int stateNumber, String stateHash){
+	public HistoryMAC(long stateNumber, String stateHash){
 		this.stateNumber = stateNumber;
 		this.stateHash = stateHash;
 	}
